[AIRLFOW-YYY] Move kerberos command to separate file
